% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BeesAndButterfliesEML}
\alias{BeesAndButterfliesEML}
\title{BeesAndButterflies EML}
\format{
An eml file
\describe{
  \item{datset}{metadata about the dataset}
  \item{additionalMetadata}{additional metadata}
  ...
}
}
\source{
\url{https://doi.org/10.15468/mpsa4g}
}
\usage{
BeesAndButterfliesEML
}
\description{
An eml document (ecological metadata) for the Bumblebees and butterflies in Norway dataset
}
\keyword{datasets}
