% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BeesAndButterfliesEvent}
\alias{BeesAndButterfliesEvent}
\title{Bumblebees and butterflies in Norway [event data]}
\format{
A data frame with 27915 rows and 26 variables:
\describe{
  \item{\code{id}}{character COLUMN_DESCRIPTION}
  \item{\code{type}}{character COLUMN_DESCRIPTION}
  \item{\code{modified}}{character COLUMN_DESCRIPTION}
  \item{\code{ownerInstitutionCode}}{character COLUMN_DESCRIPTION}
  \item{\code{dynamicProperties}}{character COLUMN_DESCRIPTION}
  \item{\code{eventID}}{character COLUMN_DESCRIPTION}
  \item{\code{parentEventID}}{character COLUMN_DESCRIPTION}
  \item{\code{samplingProtocol}}{character COLUMN_DESCRIPTION}
  \item{\code{sampleSizeValue}}{integer COLUMN_DESCRIPTION}
  \item{\code{sampleSizeUnit}}{character COLUMN_DESCRIPTION}
  \item{\code{eventDate}}{character COLUMN_DESCRIPTION}
  \item{\code{eventTime}}{character COLUMN_DESCRIPTION}
  \item{\code{eventRemarks}}{character COLUMN_DESCRIPTION}
  \item{\code{locationID}}{character COLUMN_DESCRIPTION}
  \item{\code{country}}{character COLUMN_DESCRIPTION}
  \item{\code{countryCode}}{character COLUMN_DESCRIPTION}
  \item{\code{stateProvince}}{character COLUMN_DESCRIPTION}
  \item{\code{municipality}}{character COLUMN_DESCRIPTION}
  \item{\code{locality}}{character COLUMN_DESCRIPTION}
  \item{\code{locationRemarks}}{character COLUMN_DESCRIPTION}
  \item{\code{decimalLatitude}}{double COLUMN_DESCRIPTION}
  \item{\code{decimalLongitude}}{double COLUMN_DESCRIPTION}
  \item{\code{geodeticDatum}}{character COLUMN_DESCRIPTION}
  \item{\code{coordinateUncertaintyInMeters}}{integer COLUMN_DESCRIPTION}
  \item{\code{footprintWKT}}{character COLUMN_DESCRIPTION}
  \item{\code{footprintSRS}}{character COLUMN_DESCRIPTION}
}
}
\source{
\url{https://doi.org/10.15468/mpsa4g}
}
\usage{
BeesAndButterfliesEvent
}
\description{
The Norwegian Institute for Nature Research (www.nina.no) has conducted area representative surveys of butterflies and bumblebees since 2009, on behalf of the Norwegian Environment Agency (https://www.miljodirektoratet.no/). The monitoring project is designed to provide indicators of bumblebees and butterflies to the Nature Index of Norway (https://naturindeks.no/), which measures the condition of the biodiversity in Norway. The monitoring project is supervised by the Norwegian Institute for Nature Research, but the field inventories are done by citizen scientists administered by Sabima (https://www.sabima.no/). More information (in Norwegian) can be found at the project web-page, which also has a rudimentary data display of the projects results (https://www.nina.no/Våre-fagområder/Miljøovervåking-på-land/Humler-og-dagsommerfugler ).
The project started in 2009 in the former counties Østfold (now part of county Viken) and Vestfold (now part of county Vestfold and Telemark). Citizen scientists joined the project in 2010, and the project was extended geographically the following year to include also the county Trøndelag. Since 2013, the project also includes the former county of Vest-Agder (now part of county Agder), and county Rogaland.
The surveys are currently performed at a total of 52 sites from the Lucas-grid (country covering grid network with 18 km distance between grids) in the lower parts of the regions (i.e. excluding alpine areas). The grid network is made up of square polygons, placed 18 x 18 km apart, where every square is 1.5 * 1.5 km. In each square, a total of 1 km transects (20 transects á 50 m) are placed in suitable environments (approximately evenly distributed between the habitat types open forest- and grassland), where inventories of butterflies and bumblebees are perfomed three times each summer following a standardized protocol. This includes visual identification and sweep netting along the fixed transects.
}
\keyword{datasets}
