% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FeFoEML}
\alias{FeFoEML}
\title{FeFoEML}
\format{
An eml file
\describe{
  \item{datset}{metadata about the dataset}
  \item{additionalMetadata}{additional metadata}
  ...
}
}
\source{
\url{https://doi.org/10.15468/s7c8qd}
}
\usage{
FeFoEML
}
\description{
An eml document (ecological metadata) for the Tetraonid line transect surveys from Norway: Data from Finnmarkseiendommen (FeFo)
}
\references{
{Nilsen E B, Vang R, Asbjørnsen E (2021). Tetraonid line transect surveys from Norway: Data from Finnmarkseiendommen (FeFo). Version 1.6. Norwegian Institute for Nature Research. Sampling event dataset https://doi.org/10.15468/s7c8qd accessed via GBIF.org on 2021-07-06.}
}
\keyword{datasets}
