% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FeFoOcc}
\alias{FeFoOcc}
\title{Tetraonid line transect surveys from Norway: Data from Finnmarkseiendommen (FeFo) [occurence data]}
\format{
A data frame with 8277 rows and 17 variables:
\describe{
  \item{\code{id}}{character COLUMN_DESCRIPTION}
  \item{\code{modified}}{double COLUMN_DESCRIPTION}
  \item{\code{basisOfRecord}}{character COLUMN_DESCRIPTION}
  \item{\code{occurrenceID}}{character COLUMN_DESCRIPTION}
  \item{\code{individualCount}}{double COLUMN_DESCRIPTION}
  \item{\code{sex}}{character COLUMN_DESCRIPTION}
  \item{\code{lifeStage}}{character COLUMN_DESCRIPTION}
  \item{\code{eventID}}{character COLUMN_DESCRIPTION}
  \item{\code{taxonID}}{character COLUMN_DESCRIPTION}
  \item{\code{scientificName}}{character COLUMN_DESCRIPTION}
  \item{\code{kingdom}}{character COLUMN_DESCRIPTION}
  \item{\code{phylum}}{character COLUMN_DESCRIPTION}
  \item{\code{class}}{character COLUMN_DESCRIPTION}
  \item{\code{order}}{character COLUMN_DESCRIPTION}
  \item{\code{family}}{character COLUMN_DESCRIPTION}
  \item{\code{genus}}{character COLUMN_DESCRIPTION}
  \item{\code{specificEpithet}}{character COLUMN_DESCRIPTION}
}
}
\source{
\url{https://doi.org/10.15468/s7c8qd}
}
\usage{
FeFoOcc
}
\description{
The dataset originates from a line transect survey program in Finnmark county in northern Norway carried out from 2000- and onward. The surveys are part of a national program - Hønsefuglportalen (http://honsefugl.nina.no), operated in close collaboration among NINA (www.nina.no), Nord University (www.nord.no), Inland Norway University of Applied Sciences (www.inn.no) as well as landowners responsible for grouse management on their properties (state-owned land on Finnmarkseiendommen (FeFo), Statskog, and Fjellstyrene, as well as several private landowners).
In general, NINA offers common e-infrastructure for data entry, data curation, management of users and communication from the project (http://honsefugl.nina.no). NINA together with the involved universities are offering advice regarding sampling design, and has developed a standard field protocol used by all participating stakeholders. The specific part of the program carried out on FeFo's property is operated in collaboration between NINA and FeFo, with clear and distinct allocation of tasks and responsibilities between the parties. FeFo is responsible for conducting field work, as well as final decisions regarding allocation of transect lines among study areas.
Surveys are carried out mainly by volunteer personnel instructed to follow standardized field protocols. Field work is carried out annually in August, with a team of two persons following pre-defined transect lines, using trained pointing dogs to search both sides of the transect line. When birds are observed, species, cluster size (separated into age- and sex classes), perpendicular distances from the transect line to the observed cluster of birds, GPS coordinates of the clusters and time of day are recorded. In addition, length (and geographical position) of the transect line is recorded, as well as some other auxiliary data.
}
\references{
{Nilsen E B, Vang R, Asbjørnsen E (2021). Tetraonid line transect surveys from Norway: Data from Finnmarkseiendommen (FeFo). Version 1.6. Norwegian Institute for Nature Research. Sampling event dataset https://doi.org/10.15468/s7c8qd accessed via GBIF.org on 2021-07-06.}
}
\keyword{datasets}
