% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Rock_ptarmigan}
\alias{Rock_ptarmigan}
\title{Rock_ptarmigan}
\format{
A data frame with 378 rows and 24 variables:
\describe{
  \item{\code{ID}}{integer COLUMN_DESCRIPTION}
  \item{\code{Tagnr}}{integer COLUMN_DESCRIPTION}
  \item{\code{Dato}}{integer COLUMN_DESCRIPTION}
  \item{\code{Klokke}}{integer COLUMN_DESCRIPTION}
  \item{\code{Status}}{integer COLUMN_DESCRIPTION}
  \item{\code{Plottsikkerhet}}{integer COLUMN_DESCRIPTION}
  \item{\code{Årstall}}{integer COLUMN_DESCRIPTION}
  \item{\code{Lokalitet}}{integer COLUMN_DESCRIPTION}
  \item{\code{Art}}{logical COLUMN_DESCRIPTION}
  \item{\code{Kjønn}}{integer COLUMN_DESCRIPTION}
  \item{\code{Alder}}{integer COLUMN_DESCRIPTION}
  \item{\code{Vekt}}{integer COLUMN_DESCRIPTION}
  \item{\code{Bryst}}{integer COLUMN_DESCRIPTION}
  \item{\code{Ving.1}}{integer COLUMN_DESCRIPTION}
  \item{\code{Ving.2}}{integer COLUMN_DESCRIPTION}
  \item{\code{Dødsårsak}}{integer COLUMN_DESCRIPTION}
  \item{\code{X.koordinat}}{integer COLUMN_DESCRIPTION}
  \item{\code{Y.koordinat}}{integer COLUMN_DESCRIPTION}
  \item{\code{Kondisjon}}{integer COLUMN_DESCRIPTION}
  \item{\code{Høyde_stripe_kam}}{integer COLUMN_DESCRIPTION}
  \item{\code{Lengde_stripe_kam}}{integer COLUMN_DESCRIPTION}
  \item{\code{Frekvens}}{integer COLUMN_DESCRIPTION}
  \item{\code{organismID}}{integer COLUMN_DESCRIPTION}
  \item{\code{occurrenceID}}{integer COLUMN_DESCRIPTION}
}
}
\source{
\url{https://www.gbif.org/dataset/b848f1f3-3955-4725-8ad8-e711e4a9e0ac}
}
\usage{
Rock_ptarmigan
}
\description{
This data set is based on a small radio telemetry project on rock ptarmigan (Lagopus muta) in Lierne municipality, central Norway. In total, 85 rock ptarmigans were captured and radio-collared. After release, the birds were re-located with variable geographical precision (noted in the data). When mortality were detected, cause of death was as far as possible recorded.
}
\keyword{datasets}
