% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Seapop}
\alias{Seapop}
\title{Seapop}
\format{
A data frame with 20191 rows and 50 variables:
\describe{
  \item{\code{gbifID}}{double COLUMN_DESCRIPTION}
  \item{\code{datasetKey}}{character COLUMN_DESCRIPTION}
  \item{\code{occurrenceID}}{double COLUMN_DESCRIPTION}
  \item{\code{kingdom}}{character COLUMN_DESCRIPTION}
  \item{\code{phylum}}{character COLUMN_DESCRIPTION}
  \item{\code{class}}{character COLUMN_DESCRIPTION}
  \item{\code{order}}{character COLUMN_DESCRIPTION}
  \item{\code{family}}{character COLUMN_DESCRIPTION}
  \item{\code{genus}}{character COLUMN_DESCRIPTION}
  \item{\code{species}}{character COLUMN_DESCRIPTION}
  \item{\code{infraspecificEpithet}}{logical COLUMN_DESCRIPTION}
  \item{\code{taxonRank}}{character COLUMN_DESCRIPTION}
  \item{\code{scientificName}}{character COLUMN_DESCRIPTION}
  \item{\code{verbatimScientificName}}{character COLUMN_DESCRIPTION}
  \item{\code{verbatimScientificNameAuthorship}}{logical COLUMN_DESCRIPTION}
  \item{\code{countryCode}}{character COLUMN_DESCRIPTION}
  \item{\code{locality}}{character COLUMN_DESCRIPTION}
  \item{\code{stateProvince}}{character COLUMN_DESCRIPTION}
  \item{\code{occurrenceStatus}}{logical COLUMN_DESCRIPTION}
  \item{\code{individualCount}}{double COLUMN_DESCRIPTION}
  \item{\code{publishingOrgKey}}{character COLUMN_DESCRIPTION}
  \item{\code{decimalLatitude}}{double COLUMN_DESCRIPTION}
  \item{\code{decimalLongitude}}{double COLUMN_DESCRIPTION}
  \item{\code{coordinateUncertaintyInMeters}}{logical COLUMN_DESCRIPTION}
  \item{\code{coordinatePrecision}}{logical COLUMN_DESCRIPTION}
  \item{\code{elevation}}{logical COLUMN_DESCRIPTION}
  \item{\code{elevationAccuracy}}{logical COLUMN_DESCRIPTION}
  \item{\code{depth}}{logical COLUMN_DESCRIPTION}
  \item{\code{depthAccuracy}}{logical COLUMN_DESCRIPTION}
  \item{\code{eventDate}}{double COLUMN_DESCRIPTION}
  \item{\code{day}}{logical COLUMN_DESCRIPTION}
  \item{\code{month}}{logical COLUMN_DESCRIPTION}
  \item{\code{year}}{double COLUMN_DESCRIPTION}
  \item{\code{taxonKey}}{double COLUMN_DESCRIPTION}
  \item{\code{speciesKey}}{double COLUMN_DESCRIPTION}
  \item{\code{basisOfRecord}}{character COLUMN_DESCRIPTION}
  \item{\code{institutionCode}}{character COLUMN_DESCRIPTION}
  \item{\code{collectionCode}}{character COLUMN_DESCRIPTION}
  \item{\code{catalogNumber}}{double COLUMN_DESCRIPTION}
  \item{\code{recordNumber}}{logical COLUMN_DESCRIPTION}
  \item{\code{identifiedBy}}{logical COLUMN_DESCRIPTION}
  \item{\code{dateIdentified}}{logical COLUMN_DESCRIPTION}
  \item{\code{license}}{character COLUMN_DESCRIPTION}
  \item{\code{rightsHolder}}{logical COLUMN_DESCRIPTION}
  \item{\code{recordedBy}}{logical COLUMN_DESCRIPTION}
  \item{\code{typeStatus}}{logical COLUMN_DESCRIPTION}
  \item{\code{establishmentMeans}}{logical COLUMN_DESCRIPTION}
  \item{\code{lastInterpreted}}{double COLUMN_DESCRIPTION}
  \item{\code{mediaType}}{logical COLUMN_DESCRIPTION}
  \item{\code{issue}}{character COLUMN_DESCRIPTION}
}
}
\source{
\url{https://www.gbif.org/dataset/9801530f-ab2f-4913-9050-d7239d12aed0}
}
\usage{
Seapop
}
\description{
Estimated population sizes for breeding locations based on monitoring series in closest colony during breeding season.
}
\keyword{datasets}
