% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Tov_E_EML}
\alias{Tov_E_EML}
\title{TOV_E_EML}
\format{
An eml file
\describe{
  \item{datset}{metadata about the dataset}
  \item{additionalMetadata}{additional metadata}
  ...
}#'@source \url{https://doi.org/10.15468/6jmw2e}
}
\usage{
Tov_E_EML
}
\description{
An eml document (ecological metadata) for the Extensive monitoring of breeding birds (TOV-E)
}
\keyword{datasets}
