% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Tov_E_Event}
\alias{Tov_E_Event}
\title{Extensive monitoring of breeding birds (TOV-E) [event data]}
\format{
A data frame with 3798 rows and 29 variables:
\describe{
  \item{\code{id}}{character COLUMN_DESCRIPTION}
  \item{\code{type}}{character COLUMN_DESCRIPTION}
  \item{\code{modified}}{double COLUMN_DESCRIPTION}
  \item{\code{datasetName}}{character COLUMN_DESCRIPTION}
  \item{\code{ownerInstitutionCode}}{character COLUMN_DESCRIPTION}
  \item{\code{informationWithheld}}{character COLUMN_DESCRIPTION}
  \item{\code{dataGeneralizations}}{character COLUMN_DESCRIPTION}
  \item{\code{eventID}}{character COLUMN_DESCRIPTION}
  \item{\code{samplingProtocol}}{character COLUMN_DESCRIPTION}
  \item{\code{sampleSizeValue}}{double COLUMN_DESCRIPTION}
  \item{\code{sampleSizeUnit}}{character COLUMN_DESCRIPTION}
  \item{\code{samplingEffort}}{character COLUMN_DESCRIPTION}
  \item{\code{eventDate}}{double COLUMN_DESCRIPTION}
  \item{\code{eventTime}}{character COLUMN_DESCRIPTION}
  \item{\code{year}}{double COLUMN_DESCRIPTION}
  \item{\code{month}}{double COLUMN_DESCRIPTION}
  \item{\code{day}}{double COLUMN_DESCRIPTION}
  \item{\code{locationID}}{double COLUMN_DESCRIPTION}
  \item{\code{country}}{character COLUMN_DESCRIPTION}
  \item{\code{countryCode}}{character COLUMN_DESCRIPTION}
  \item{\code{stateProvince}}{character COLUMN_DESCRIPTION}
  \item{\code{municipality}}{character COLUMN_DESCRIPTION}
  \item{\code{locality}}{character COLUMN_DESCRIPTION}
  \item{\code{minimumElevationInMeters}}{double COLUMN_DESCRIPTION}
  \item{\code{maximumElevationInMeters}}{double COLUMN_DESCRIPTION}
  \item{\code{decimalLatitude}}{double COLUMN_DESCRIPTION}
  \item{\code{decimalLongitude}}{double COLUMN_DESCRIPTION}
  \item{\code{geodeticDatum}}{character COLUMN_DESCRIPTION}
  \item{\code{coordinateUncertaintyInMeters}}{double COLUMN_DESCRIPTION}
}
}
\source{
\url{https://doi.org/10.15468/6jmw2e}
}
\usage{
Tov_E_Event
}
\description{
Data from the project Extensive monitoring of breeding birds (TOV-E); from 2006 up until today. The project is carried out in cooperation between NOF BirdLife Norway, Norwegian Institute for Nature Research (NINA) and the Norwegian Environment Agency, and is the most important project for monitoring population trends for Norwegian bird species on land
}
\keyword{datasets}
