% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Tov_EOcc}
\alias{Tov_EOcc}
\title{Extensive monitoring of breeding birds (TOV-E) [occurrence data]}
\format{
A data frame with 85759 rows and 20 variables:
\describe{
  \item{\code{id}}{character COLUMN_DESCRIPTION}
  \item{\code{type}}{character COLUMN_DESCRIPTION}
  \item{\code{modified}}{double COLUMN_DESCRIPTION}
  \item{\code{collectionCode}}{character COLUMN_DESCRIPTION}
  \item{\code{basisOfRecord}}{character COLUMN_DESCRIPTION}
  \item{\code{occurrenceID}}{character COLUMN_DESCRIPTION}
  \item{\code{catalogNumber}}{double COLUMN_DESCRIPTION}
  \item{\code{occurrenceRemarks}}{character COLUMN_DESCRIPTION}
  \item{\code{organismQuantity}}{double COLUMN_DESCRIPTION}
  \item{\code{organismQuantityType}}{character COLUMN_DESCRIPTION}
  \item{\code{eventID}}{character COLUMN_DESCRIPTION}
  \item{\code{scientificName}}{character COLUMN_DESCRIPTION}
  \item{\code{kingdom}}{character COLUMN_DESCRIPTION}
  \item{\code{phylum}}{character COLUMN_DESCRIPTION}
  \item{\code{class}}{character COLUMN_DESCRIPTION}
  \item{\code{order}}{character COLUMN_DESCRIPTION}
  \item{\code{family}}{character COLUMN_DESCRIPTION}
  \item{\code{genus}}{character COLUMN_DESCRIPTION}
  \item{\code{infraspecificEpithet}}{character COLUMN_DESCRIPTION}
  \item{\code{vernacularName}}{character COLUMN_DESCRIPTION}
}
}
\source{
\url{https://doi.org/10.15468/6jmw2e}
}
\usage{
Tov_EOcc
}
\description{
Data from the project Extensive monitoring of breeding birds (TOV-E); from 2006 up until today. The project is carried out in cooperation between NOF BirdLife Norway, Norwegian Institute for Nature Research (NINA) and the Norwegian Environment Agency, and is the most important project for monitoring population trends for Norwegian bird species on land
}
\keyword{datasets}
